pyproject.toml
[project]
name = "repository-intelligence-scanner"
version = "1.1.0"
description = "Decision-grade repository analysis system"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
]
api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "aiofiles>=23.2.1",
]
ai = [
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "llama-cpp-python>=0.2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
repo-scanner = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "tests*"]
